+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - 當high-concurrency-processor調用下游Server時，遭遇readTimeout的處理
+- 新增子專案system-gateway
 
 ## [0.8.1] - 2024-03-12